requirements.txt
openai>=1.42.0
pydantic>=2.8.2
azure-identity>=1.18.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
